add: data.js and util.js, fix: finding id and url
